+++ bears/xml2/XMLBear.py
