+++ tests/xml2/test_files/dtd-error.xml
