+++ tests/xml2/test_files/note.xsd
