+++ tests/xml2/test_files/xsd-error.xml
